fix: all printk fmt related with %llx

+++ c_depot/early_dt/src/init.c
@@ -451,8 +451,8 @@ early_init_dt_scan_memory(unsigned long node,
             continue;
 
         printk(" - %lx ,  %lx\n",
-               (unsigned long long)base,
-               (unsigned long long)size);
+               (unsigned long)base,
+               (unsigned long)size);
 
         early_init_dt_add_memory_arch(base, size);
     }
